update Writer Visitor docs, move Config to sep file
